# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: forza-telemetry-service
  version: v1
servers:
- url: /
paths:
  /api/v1/live/info:
    get:
      operationId: info
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LiveInfoDTO"
          description: OK
      tags:
      - live-controller
  /api/v1/live/snapshot:
    get:
      operationId: snapshot
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LiveSnapshotDTO"
          description: Último pacote recebido
        "404":
          content:
            application/problem+json: {}
          description: Nenhum pacote recente
      tags:
      - live-controller
  /api/v1/sessions:
    get:
      operationId: list
      parameters:
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SessionPageDTO"
          description: Página de sessões
        "400":
          content:
            application/problem+json: {}
          description: Cursor inválido
      tags:
      - session-controller
  /api/v1/sessions/{id}:
    get:
      operationId: findById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SessionDTO"
          description: Metadados da sessão
        "404":
          content:
            application/problem+json: {}
          description: Sessão não encontrada
      tags:
      - session-controller
  /api/v1/sessions/{id}/laps:
    get:
      operationId: laps
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LapDTO"
          description: OK
      tags:
      - session-controller
  /api/v1/sessions/{id}/samples:
    get:
      operationId: samples
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: fromMs
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - in: query
        name: toMs
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 2000
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SampleDTO"
          description: OK
      tags:
      - session-controller
  /api/v1/sessions/{id}/summary:
    get:
      operationId: summary
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TuningSummaryDTO"
          description: Resumo de tuning
        "404":
          content:
            application/problem+json: {}
          description: Sessão não encontrada
      tags:
      - session-controller
components:
  schemas:
    Braking:
      type: object
      properties:
        frontLockPct:
          type: number
          format: double
        rearLockPct:
          type: number
          format: double
        samples:
          type: integer
          format: int32
    CornerBalance:
      type: object
      properties:
        entry:
          $ref: "#/components/schemas/CornerPhase"
        exit:
          $ref: "#/components/schemas/CornerPhase"
        mid:
          $ref: "#/components/schemas/CornerPhase"
        note:
          type: string
    CornerPhase:
      type: object
      properties:
        frontSlipAngleMean:
          type: number
          format: double
        oversteerPct:
          type: number
          format: double
        rearSlipAngleMean:
          type: number
          format: double
        samples:
          type: integer
          format: int32
        understeerPct:
          type: number
          format: double
    Engine:
      type: object
      properties:
        boostMaxPsi:
          type: number
          format: double
        gears:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Gear"
        maxRpm:
          type: number
          format: double
        peakPowerHp:
          type: number
          format: double
        peakPowerRpm:
          type: number
          format: double
        peakTorqueNm:
          type: number
          format: double
        peakTorqueRpm:
          type: number
          format: double
    Gear:
      type: object
      properties:
        limiterWithThrottlePct:
          type: number
          format: double
        rpmP50:
          type: number
          format: double
        timePct:
          type: number
          format: double
    Lap:
      type: object
      properties:
        lap:
          type: integer
          format: int32
        timeS:
          type: number
          format: double
    LapDTO:
      type: object
      properties:
        lapNumber:
          type: integer
          format: int32
        lapTimeS:
          type: number
          format: float
    Laps:
      type: object
      properties:
        bestS:
          type: number
          format: double
        times:
          type: array
          items:
            $ref: "#/components/schemas/Lap"
    LiveInfoDTO:
      type: object
      properties:
        hostAddresses:
          type: array
          items:
            type: string
        udpPort:
          type: integer
          format: int32
    LiveSnapshotDTO:
      type: object
      properties:
        accel:
          type: integer
          format: int32
        bestLapS:
          type: number
          format: float
        brake:
          type: integer
          format: int32
        carName:
          type: string
        carOrdinal:
          type: integer
          format: int32
        combinedSlip:
          type: array
          items:
            type: number
            format: float
        currentLapS:
          type: number
          format: float
        engineMaxRpm:
          type: number
          format: float
        gameFormat:
          type: string
        gear:
          type: integer
          format: int32
        lapNumber:
          type: integer
          format: int32
        lastLapS:
          type: number
          format: float
        performanceIndex:
          type: integer
          format: int32
        raceOn:
          type: boolean
        receivedAt:
          type: string
          format: date-time
        rpm:
          type: number
          format: float
        slipAngle:
          type: array
          items:
            type: number
            format: float
        speedKmh:
          type: number
          format: float
        steer:
          type: integer
          format: int32
        suspension:
          type: array
          items:
            type: number
            format: float
        tireTempF:
          type: array
          items:
            type: number
            format: float
    SampleDTO:
      type: object
      properties:
        accel:
          type: integer
          format: int32
        accelX:
          type: number
          format: float
        accelZ:
          type: number
          format: float
        boost:
          type: number
          format: float
        brake:
          type: integer
          format: int32
        combinedSlip:
          type: array
          items:
            type: number
            format: float
        gear:
          type: integer
          format: int32
        lapNumber:
          type: integer
          format: int32
        onRumble:
          type: boolean
        posX:
          type: number
          format: float
        posZ:
          type: number
          format: float
        power:
          type: number
          format: float
        rpm:
          type: number
          format: float
        slipAngle:
          type: array
          items:
            type: number
            format: float
        slipRatio:
          type: array
          items:
            type: number
            format: float
        speed:
          type: number
          format: float
        steer:
          type: integer
          format: int32
        suspension:
          type: array
          items:
            type: number
            format: float
        tMs:
          type: integer
          format: int32
        tireTemp:
          type: array
          items:
            type: number
            format: float
        tireWear:
          type: array
          items:
            type: number
            format: float
        torque:
          type: number
          format: float
    SessionDTO:
      type: object
      properties:
        active:
          type: boolean
        carClass:
          type: integer
          format: int32
        carName:
          type: string
        carOrdinal:
          type: integer
          format: int32
        cylinders:
          type: integer
          format: int32
        drivetrain:
          type: string
        endedAt:
          type: string
          format: date-time
        gameFormat:
          type: string
        id:
          type: string
          format: uuid
        performanceIndex:
          type: integer
          format: int32
        sampleCount:
          type: integer
          format: int32
        startedAt:
          type: string
          format: date-time
        trackOrdinal:
          type: integer
          format: int32
    SessionPageDTO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SessionDTO"
        nextCursor:
          type: string
    SpeedKmh:
      type: object
      properties:
        max:
          type: number
          format: double
        meanMoving:
          type: number
          format: double
    SuspensionWheel:
      type: object
      properties:
        bottomingPct:
          type: number
          format: double
        mean:
          type: number
          format: double
        p95:
          type: number
          format: double
        toppingPct:
          type: number
          format: double
    TireWheel:
      type: object
      properties:
        tempMeanF:
          type: number
          format: double
        tempP95F:
          type: number
          format: double
        wearFinal:
          type: number
          format: double
    Traction:
      type: object
      properties:
        drivenWheelSpinPct:
          type: number
          format: double
        samples:
          type: integer
          format: int32
        spinPctByGear:
          type: object
          additionalProperties:
            type: number
            format: double
    TuningSummaryDTO:
      type: object
      description: Resumo de tuning de uma sessão
      properties:
        braking:
          $ref: "#/components/schemas/Braking"
        cornerBalance:
          $ref: "#/components/schemas/CornerBalance"
        durationS:
          type: number
          format: double
        engine:
          $ref: "#/components/schemas/Engine"
        laps:
          $ref: "#/components/schemas/Laps"
        onRumbleStripPct:
          type: number
          format: double
        samples:
          type: integer
          format: int32
        speedKmh:
          $ref: "#/components/schemas/SpeedKmh"
        suspension:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/SuspensionWheel"
        tires:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/TireWheel"
        traction:
          $ref: "#/components/schemas/Traction"